// Angular/src/swagger-generated/model/product.ts
/**
 * Api Documentation
 * Api Documentation
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { User } from './user';


export interface Product { 
    code?: string;
    description?: string;
    id?: number;
    price?: string;
    quantity?: number;
    user?: User;
}